# Conflicts:
#	src/main/main.js

+++ src/main/main.js
@@ -70,9 +70,9 @@ async function createPluginControllerWindow({
   pluginControllerWindow.removeMenu();
 
   await pluginControllerWindow.loadURL(path.join(baseUrl, 'public', 'plugin.html'));
-  if (process.env.NODE_ENV === 'development') {
+  // if (process.env.NODE_ENV === 'development') {
     pluginControllerWindow.webContents.openDevTools();
-  }
+  // }
 
   pluginControllerWindow.webContents.send('new-plugin-loaded', {
     pluginKey,
@@ -224,13 +224,12 @@ async function createWindow() {
 
   mainWindow.loadURL(path.join(baseUrl, 'public', 'index.html'));
 
-  const defaultRatio = 4 / 3;
-
-  if (process.platform === 'darwin') {
-    mainWindow.setAspectRatio(defaultRatio);
-  } else {
+  if (process.platform === 'win32') {
     mainWindowHandler = new aspect(mainWindow);
-    mainWindowHandler.setRatio(4, 3, 100);
+    mainWindowHandler.setRatio(4, 3, 10);
+  } else {
+    const defaultRatio = 4 / 3;
+    mainWindow.setAspectRatio(defaultRatio);
   }
 
   if (process.env.NODE_ENV === 'development') {
@@ -336,13 +335,14 @@ async function createWindow() {
       });
 
       await new Promise((resolve, reject) => {
-        ipcMain.once('cancel-plugin-installation', async () => {
-          log('installation cancelled by user');
-          store.delete(`pluginData.${pluginKey}`);
-          reject(new Error('Installation Cancelled'));
-        });
-        ipcMain.once('install-plugin', async () => {
-          resolve();
+        ipcMain.once('continue-plugin-installation', (e, { shouldContinue }) => {
+          if (shouldContinue) {
+            resolve();
+          } else {
+            log('installation cancelled by user');
+            store.delete(`pluginData.${pluginKey}`);
+            reject(new Error('Installation Cancelled'));
+          }
         });
       });
 
@@ -393,9 +393,8 @@ async function createWindow() {
         icon,
         pluginSettings,
         acceptRestrictions,
-        isLoading: false,
         isInstalling: false,
-        isWorking: 'idle',
+        isWorking: false,
       });
       return manifestData;
     } catch (err) {
@@ -405,9 +404,9 @@ async function createWindow() {
     }
   };
 
-  ipcMain.on('recieved-plugin-tarball', async (event, filesArr) => {
+  ipcMain.on('recieved-plugin-tarball', async (event, filePath) => {
     try {
-      const result = await unpackPlugin(filesArr[0].path);
+      const result = await unpackPlugin(filePath);
       if (result) {
         mainWindow.webContents.send('manifest-changed', 'add');
       }
@@ -533,43 +532,58 @@ ipcMain.on('restart-app', () => {
   app.relaunch();
 });
 
-app.on('activate', () => {
-  if (mainWindow === null) createWindow();
-});
+const isSingleAppInstance = app.requestSingleInstanceLock();
 
-app.on('ready', () => {
-  log('App ready, creating window...');
-
-  createWindow();
-
-  log('Created main-renderer window, registering protocol...');
-
-  protocol.registerStringProtocol('deskfiler', (request, callback) => {
-    log('Registered deskfiler:// protocol');
-
-    callback();
+if (!isSingleAppInstance) {
+  app.quit();
+} else {
+  app.on('second-instance', () => {
+    if (mainWindow) {
+      if (mainWindow.isMinimized()) mainWindow.restore();
+      mainWindow.focus();
+    }
   });
 
-  log('Initializing serve for plugins');
-
-  server = http.createServer((request, response) => (
-    handler(request, response, {
-      public: PLUGINS_DIR,
-    })
-  ));
-
-  server.listen(PORT, () => {
-    log(`Hosting plugins @ http://localhost:${PORT}.`);
+  app.on('activate', () => {
+    if (mainWindow === null) createWindow();
   });
-});
+  
+  app.on('ready', () => {
+    log('App ready, creating window...');
+  
+    createWindow();
+  
+    log('Created main-renderer window, registering protocol...');
+  
+    protocol.registerStringProtocol('deskfiler', (request, callback) => {
+      log('Registered deskfiler:// protocol');
+  
+      callback();
+    });
+  
+    log('Initializing serve for plugins');
+  
+    server = http.createServer((request, response) => (
+      handler(request, response, {
+        public: PLUGINS_DIR,
+      })
+    ));
+  
+    server.listen(PORT, () => {
+      log(`Hosting plugins @ http://localhost:${PORT}.`);
+    });
+  });
+  
+  app.on('login', (event, webContents, request, authInfo, callback) => {
+    event.preventDefault();
+    callback('a', 'b');
+  });
+  
+  app.on('window-all-closed', () => {
+    server.close(() => { log('Plugins server closed.'); });
+    log('Terminating app...');
+    if (process.platform !== 'darwin') app.quit();
+  });
+};
 
-app.on('login', (event, webContents, request, authInfo, callback) => {
-  event.preventDefault();
-  callback('a', 'b');
-});
 
-app.on('window-all-closed', () => {
-  server.close(() => { log('Plugins server closed.'); });
-  log('Terminating app...');
-  if (process.platform !== 'darwin') app.quit();
-});
